% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{all_genes}
\alias{all_genes}
\title{get all known O.sative gene ids - the universe
genefile csv file with GeneID column of all O. sativa ids}
\usage{
all_genes(which = "indica")
}
\arguments{
\item{which}{subspecies with either "indica" or "japonica". Default is "indica"}
}
\description{
get all known O.sative gene ids - the universe
genefile csv file with GeneID column of all O. sativa ids
}
