% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{all_mo_genes}
\alias{all_mo_genes}
\title{get all known M.oryzae gene ids - the universe}
\usage{
all_mo_genes(genefile = "inst/extdata/all_mo_ids.csv")
}
\arguments{
\item{genefile}{csv file with GeneID column of all Mo ids}
}
\description{
get all known M.oryzae gene ids - the universe
}
