% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{do_enrich}
\alias{do_enrich}
\title{run clusterProfiler::enricher on vector of gene ids,}
\usage{
do_enrich(genes, termsfile = "inst/extdata/MG8_mart_export.txt", ...)
}
\arguments{
\item{genes}{character vector of gene IDs of interest}

\item{termsfile}{TSV file from biomart export of GO terms}
}
\value{
enricher object
}
\description{
run clusterProfiler::enricher on vector of gene ids,
}
