% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{enricher_to_david}
\alias{enricher_to_david}
\title{convert enricher object to DAVID format}
\usage{
enricher_to_david(enrich, which = "indica")
}
\arguments{
\item{enrich}{enricher object from do_enrich}

\item{which}{subspecies with either "indica" or "japonica". Default is "indica"}
}
\value{
DAVID format dataframe
}
\description{
convert enricher object to DAVID format
}
