% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mapping}
\alias{mapping}
\title{get mapping between Os Gene IDs and GO Terms}
\usage{
mapping(termsfile = NULL, which = "indica")
}
\arguments{
\item{termsfile}{TSV file from export of GO terms for indica group at plants.ensembl.org Oryza sativa Japonica Group genes (IRGSP-1.0)}

\item{which}{subspecies with either "indica" or "japonica". Default is "indica"}
}
\description{
get mapping between Os Gene IDs and GO Terms
}
