% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mo_mapping}
\alias{mo_mapping}
\title{get mapping between Mo Gene IDs and GO Terms}
\usage{
mo_mapping(termsfile = "inst/extdata/MG8_mart_export.txt")
}
\arguments{
\item{termsfile}{TSV file from biomart export of GO terms}
}
\description{
get mapping between Mo Gene IDs and GO Terms
}
