% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{read_terms}
\alias{read_terms}
\title{read the terms file}
\usage{
read_terms(termsfile = NULL)
}
\arguments{
\item{termsfile}{TSV file from export of GO terms for indica group at plants.ensembl.org Oryza sativa Japonica Group genes (IRGSP-1.0)}
}
\description{
read the terms file
}
